finalized all files for the solutions of the knapsack problem

+++ assignment_2/knapsack_problem/solution_algorithms/_knapsack_types.py
@@ -3,7 +3,7 @@ from collections import namedtuple
 
 Item = namedtuple('Item', ['name', 'weight', 'value', 'n_items'])
 Items = List[Item]
-WeightLimit = int
+WeightLimit = float
 
 # Genetic Algorithm Types
 Gene = int
